use userEmailContext during tool execution stage
