+++ front/lib/actions/utils.ts
@@ -1,7 +1,6 @@
 import { ToolsIcon } from "@dust-tt/sparkle";
 
 import type { ActionSpecification } from "@app/components/assistant_builder/types";
-import type { MCPToolStakeLevelType } from "@app/lib/actions/constants";
 import type { MCPToolConfigurationType } from "@app/lib/actions/mcp";
 import type { StepContext } from "@app/lib/actions/types";
 import {
@@ -14,10 +13,7 @@ import {
   isMCPInternalWebsearch,
 } from "@app/lib/actions/types/guards";
 import { getSupportedModelConfig } from "@app/lib/assistant";
-import type { Authenticator } from "@app/lib/auth";
-import type { UserResource } from "@app/lib/resources/user_resource";
-import type { AgentConfigurationType, AgentMessageType } from "@app/types";
-import { assertNever } from "@app/types";
+import type { AgentConfigurationType } from "@app/types";
 
 export const WEBSEARCH_ACTION_NUM_RESULTS = 16;
 export const SLACK_SEARCH_ACTION_NUM_RESULTS = 24;
@@ -176,106 +172,4 @@ export function computeStepContexts({
   }
 
   return stepContexts;
-}
-
-export async function getExecutionStatusFromConfig(
-  auth: Authenticator,
-  actionConfiguration: MCPToolConfigurationType,
-  agentMessage: AgentMessageType
-): Promise<{
-  stake?: MCPToolStakeLevelType;
-  status: "ready_allowed_implicitly" | "blocked_validation_required";
-  serverId?: string;
-}> {
-  // If the agent message is marked as "skipToolsValidation" we skip all tools validation
-  // irrespective of the `actionConfiguration.permission`. This is set when the agent message was
-  // created by an API call where the caller explicitly set `skipToolsValidation` to true.
-  if (agentMessage.skipToolsValidation) {
-    return { status: "ready_allowed_implicitly" };
-  }
-
-  // Permissions:
-  // - "never_ask": Automatically approved
-  // - "low": Ask user for approval and allow to automatically approve next time
-  // - "high": Ask for approval each time
-  // - undefined: Use default permission ("never_ask" for default tools, "high" for other tools)
-  switch (actionConfiguration.permission) {
-    case "never_ask":
-      return { status: "ready_allowed_implicitly" };
-    case "low": {
-      // The user may not be populated, notably when using the public API.
-      const user = auth.user();
-
-      if (
-        user &&
-        (await hasUserAlwaysApprovedTool({
-          user,
-          mcpServerId: actionConfiguration.toolServerId,
-          functionCallName: actionConfiguration.name,
-        }))
-      ) {
-        return { status: "ready_allowed_implicitly" };
-      }
-      return { status: "blocked_validation_required" };
-    }
-    case "high":
-      return { status: "blocked_validation_required" };
-    default:
-      assertNever(actionConfiguration.permission);
-  }
-}
-
-const TOOLS_VALIDATION_WILDCARD = "*";
-
-const getToolsValidationKey = (mcpServerId: string) =>
-  `toolsValidations:${mcpServerId}`;
-
-// The function call name is scoped by MCP servers so that the same tool name on different servers
-// does not conflict, which is why we use it here instead of the tool name.
-export async function setUserAlwaysApprovedTool({
-  user,
-  mcpServerId,
-  functionCallName,
-}: {
-  user: UserResource;
-  mcpServerId: string;
-  functionCallName: string;
-}) {
-  if (!functionCallName) {
-    throw new Error("functionCallName is required");
-  }
-  if (!mcpServerId) {
-    throw new Error("mcpServerId is required");
-  }
-
-  await user.upsertMetadataArray(
-    getToolsValidationKey(mcpServerId),
-    functionCallName
-  );
-}
-
-export async function hasUserAlwaysApprovedTool({
-  user,
-  mcpServerId,
-  functionCallName,
-}: {
-  user: UserResource;
-  mcpServerId: string;
-  functionCallName: string;
-}) {
-  if (!mcpServerId) {
-    throw new Error("mcpServerId is required");
-  }
-
-  if (!functionCallName) {
-    throw new Error("functionCallName is required");
-  }
-
-  const metadata = await user.getMetadataAsArray(
-    getToolsValidationKey(mcpServerId)
-  );
-  return (
-    metadata.includes(functionCallName) ||
-    metadata.includes(TOOLS_VALIDATION_WILDCARD)
-  );
 }
